Changing Approach
Ran into a brick wall with various existing SOAP libraries. None allowed me to modify the envelope, and they ran into infinite loop errors when retrieving the test wsdl from MS.

Changing the approach to just construct the envelopes manually from a .txt template and use the requests package to send it.

This is now working, but returns a 404. I've temporarily overridden the env vars with the example data given in the API docs and it's evident that the hash of the signature is incorrect.

I think this template approach will be successful. Now troubleshooting the signature hashing (they give you this test data and say what the resulting hash should be to confirm).