Mod ConciergeClient API
Take secrets as arguments rather than pulling directly from the
environment.

This isn't tested and the new method, login_to_portal(), is
psuedo-code.  I'm committing the changes so we can discuss them in the
context of a pull request.  Is this a good idea?  If we agree it is,
I'll continue to refine the code and adjust the tests.