Add ability to filter by last modified
Added positional argument to query_orgs() to filter by when an organization was last modified.

Added additional exception catching to the function as well to handle returning null sets, and necessary logic for if you have one, both, or neither of the positional arguments present.

Updated test to attempt to query orgs modified in the last day. This test will likely break at some point in the future until we set up a sandbox to point this at instead (where it should hold true that we haven't modified any objects for at least a day).