Add Object Conversion Method
Added generic method to turn the list of attributes of an object returned from an API method into a dictionary.

Some attributes will still be complex and require further processing, but that can be localized (or you can call this same method again and pass just that field as the argument).