Recursion Fix
Recursion was working fine for testing because max_depth was an integer.

In a "real world" case where the default of None is used instead, the
comparison of depth < max_depth is always False so recursion does not
happen. Changed to instead test for inequality, so if it's None that is
always True instead.