Add `delay` argument to security.services.login_to_portal()
When `retries` argument is > 0, login_to_portal() will retry logging
in if it fails, waiting `delay` seconds between retries.

`delay` defaults to 0 so usually no pause will occur.  It's for tests,
to try to avoid potential race conditions.  In the security tests,
every call to login_to_portal() sets retries to 5 and delay to 1.
Maybe that'll work.