Mixin Fix
Mixin that rewrote how queries are batched didn't take into account the full
structure of the api's response object. len(result) was always 2 becuase it
contains a little status dictionary and then the actual results set nested object
tree.

Needed to dig deeper and take the length of the innermost item that is actually
the list of objects.